Edited file 'python-lib/sage/src/dss_init.py'

+++ python-lib/sage/src/dss_init.py
@@ -127,8 +127,9 @@ def get_dss_commits(project_handle):
 def create_scenarios(project_handle, location):
     # Clear out any old
     for scenario in project_handle.list_scenarios():
-        scenario_handle = project.get_scenario(scenario["id"])
-        scenario_handle.delete()
+        if "data_gather_" in scenario["name"]:
+            scenario_handle = project.get_scenario(scenario["id"])
+            scenario_handle.delete()
     
     # Create the scenarios
     if location ==  "WORKER":
